moving the homepage content and video components to the /components/homepage folder, it was getting a bit difficult to navigate in the homepage
